updated PE power test with new testbench

+++ mflowgen/common/cadence-xcelium-sim/run_sim.sh
@@ -3,14 +3,16 @@
 sed -i "1 i\`define CLK_PERIOD ${clock_period}" inputs/testbench.sv
 
 # Default arguments
-ARGS="-sv -timescale 1ns/1ns -access +rwc"
+ARGS="-sv -timescale 1ns/1ps -access +rwc"
 ARGS="$ARGS -ALLOWREDEFINITION"
 if [ -f "inputs/cmd.tcl" ]; then
   ARGS="$ARGS -input inputs/cmd.tcl"
 fi
 
 if [ "$PWR_AWARE" = "True" ]; then
-  rm inputs/design.vcs.v
+    if [ -f "inputs/design.vcs.v" ]; then
+        rm inputs/design.vcs.v
+    fi
 fi
 
 # ADK for GLS
